Implement y-ticks in PDE subclass
y ticks should be equation specific
